requirements.txt
gradio
ultralytics
pandas
openpyxl
Pillow
flask
pytz
uuid
